src/app/components/dashboard/dashboard.component.ts
import { Component, OnInit } from '@angular/core';
import { DataService } from '../../services/DataService';
import { OrganizationModel } from '../../models/OrganizationModel';

@Component({
  selector: 'app-dashboard',
  templateUrl: './dashboard.component.html',
  styleUrls: ['./dashboard.component.scss']
})
export class DashboardComponent implements OnInit {

  organizations: OrganizationModel[] = [];

  constructor(private readonly dataService: DataService) {}

  ngOnInit(): void {
    this.dataService.getOrganizations().subscribe((success: OrganizationModel[]) => {
      console.log('DATA:', success[0]?.name);
      this.organizations = success;
    });
  }
}
